fix: args were positioned incorrectly (#1799)
## Description:
@barnabasbusa submitted an
[issue](https://github.com/kurtosis-tech/kurtosis/issues/1789) that
demonstrated that once run, the package would not apply the settings
provided in the package `args`. I identified a bug that meant that while
we were submitting the data it was in the wrong position and needed to
be moved up on level (destructured).

In this PR we are also setting the JSON type as the default arg type.

## Is this change user facing?
YES

## References (if applicable):
Closes https://github.com/kurtosis-tech/kurtosis/issues/1789