feat: generate enclave manager ui in build process and check prettier (#1717)
## Description:
This pr includes the new enclave manager in the build process -
specifically in the `engine` build process. This removes the need to
commit the built emui to the codebase.

Additionally `yarn prettier` is added as a build step to the emui, which
will fail the build if prettier has not been run (as is currently the
case in main).

This change also includes a fix to all linting warnings, so that there
are no warnings in the production build.

## Is this change user facing?
NO

## References (if applicable):
* devconnect burndown