feat: emui auth via cookie (#1783)
## Description:
This PR modifies how the AuthenticatedKurtosisClient is created by
retrieving authentication material from a cookie, rather than via a
passed message.

This PR adds back the connect button.

Finally, this pr adds support for locally running proxied inside of the
cloud frontend behind a local nginx server. This improves the
development experience of the cloud/emui portal.

## Is this change user facing?
YES - This is also a breaking change and will require the corresponding
change in the cloud frontend to be deployed to communicate the new auth
material properly

---------

Co-authored-by: Anders Schwartz <adschwartz@users.noreply.github.com>
Co-authored-by: Anders Schwartz <anders.schwartz@kurtosistech.com>