fix: always restart logs aggregator (#1841)
## Description:
This change strengthens the restart policy for the logs aggregator.
Prior to this, the restart only occurred on failure. Now, we make docker
attempt to always restart the logs aggregator. This should help address
https://github.com/kurtosis-tech/kurtosis/issues/1832 where the logs
aggregator was stopped with a `137` status code but wasn't restarted.

This change also addresses a `Propagate must be provided with a cause`
panic occurred here:
https://github.com/kurtosis-tech/kurtosis/issues/1832. This was caused
by nil err's being propagated in the create logs collector code. This
change fixes that issue.

## Is this change user facing?
NO

## References:
https://github.com/kurtosis-tech/kurtosis/issues/1832
https://github.com/kurtosis-tech/kurtosis/issues/1311