feat: emui package details page (#1873)
## Description:
This PR implements the emui package details page.

The change @adschwartz to refactor keyboard listener code requested in
https://github.com/kurtosis-tech/kurtosis/pull/1865 is also included.

It also fixes the poor rendering of the catalog logos on chrome. 

Additionally it implements feedback Tise left on the designs for me to
pick up, specifically:
* Value card icon appearance
* Use preferred round star icon
* Update the run button appearance and interaction behaviour to match
the designs.

Finally, quite a large addition is included to support the combination
of `react-markdown` and `chakra` in the `KurtosisMarkdown` component
which includes basic chakra implementations of the html components. This
is required because `CSSReset` (part of chakra theme) unsets all of the
styling that would apply to markdown.

### Short demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/19826393-10b2-40de-832b-6970f305a2ce

## Is this change user facing?
YES

## References (if applicable):
* Figma