docs: remove cloud early access signup form (#1878)
## Description:
Cloud is real now and works amazingly. This PR replaces mentions of an
early sign-up form for Beta access to Cloud with an actual link to
cloud.kurtosis.com.

The old form was made and put in our docs to collect early feedback and
design partners in the absence of an actual product for people to try.
We now have a real cloud offering for people to try.

## Is this change user facing?
YES

## References (if applicable):
https://kurtosistech.slack.com/archives/C051H93B6DV/p1701435264644419