chore: emui artifact layout fix (#1877)
## Description:
The change to introduce optional gray headers in the background of the
emui broke the flexbox sizing for the artifact page (it was using
viewport scroll, rather than scrolls on the inner containers). This is
fixed by correcting the usage of flexbox (most specifically allowing
`min-height:0` on the app wrapper flex when page scroll is disabled -
this allows flexbox to squash children and have them deal with
overflow).

### Validation of change on every page in app.


https://github.com/kurtosis-tech/kurtosis/assets/4419574/04999b05-37f9-41f7-95b3-c92cc5048bb3

Locally I've also taken the same validation steps in chrome.

## Is this change user facing?
YES

## References (if applicable):
screenshot of broken artifact layout:

![image](https://github.com/kurtosis-tech/kurtosis/assets/4419574/ce137658-8dad-45a3-85c8-48a56932e7ec)