feat: service logs full download (#1895)
## Description:
This PR implements the full streamed download of service logs. This is
based on some work already done by @tedim52 in tedi/copydownload.

Changes:
* The copy logs button is no longer available on the Service Logs page
* The download logs button sends the log stream to a stream writer to
download them as a file.

Concerns:
* For services that log excessively, like the log spammer, it seems
downloading the logs is incredibly slow - <1kb/s. I'm not sure what
could improve this, but if the backend streamed more than one record per
message (as seems to be possible by the api spec) that seems like it
would help (and would help improve performance when rendering service
logs in the ui).

### Demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/61716034-fb6e-47c1-b776-e8c262b40799

## Is this change user facing?
YES

---------

Co-authored-by: Anders Schwartz <adschwartz@users.noreply.github.com>