feat: Reverse proxy lifecycle management and connectivity on Docker (#1906)
## Description:
In #1871 we added the Traefik labels to the user services so Traefik can
discover them and route to them. This PR implements a reverse proxy
using Traefik on Docker. The engine starts and stops Traefik. The
Traefik container is automatically connected to the enclave networks.
The implementation leverages what was done for the logs aggregator.

## Is this change user facing?
YES

## References (if applicable):
#1871