chore: emui sharable components refactor (#1923)
## Description:
This PR refactors some code out of the emui `components` directory
(mostly enclave related) and tweaks code in the `components` directory
to remove any dependencies on `KurtosisClient` (EM) and emui app code.

This change prepares the `components` directory to be refactored into a
shared components libary. The directory does still contain references to
the package indexer (which is fine for the catalog.kurtosis.com
application), which will need to be refactored later.

## Is this change user facing?
NO - no functionality is changed