chore: fix clean yarn start in emui (#1973)
## Description:
After the EMUI refactor to a multipackage build running `yarn start` on
a clean checkout failed with `Can't resolve: 'kurtosis-ui-components'`.

This failure was because the `build/` output is not present on a clean
checkout. Additionally, by referencing components `build` directory,
rather than `src`, hot reloading changes to the components in the emui
was also not working.


![image](https://github.com/kurtosis-tech/kurtosis/assets/4419574/46242777-5fbb-4cba-979b-f9a7dba073bf)

This PR fixes this issue by:
* Adding a `start` script to the components package, which hot rebuilds
on changes (via ` tsc --watch`). This enables the app to reload on
changes to the `components` source.
* Waits for the `components/build` directory to be populated before
running the development server of the `app` - this enables it to
initially load correctly.

## Is this change user facing?
NO

## References (if applicable):

https://sageroomworkspace.slack.com/archives/C05DZ4UABDZ/p1702590613774869?thread_ts=1702549439.774119&cid=C05DZ4UABDZ