fix: Make the reverse proxy connect and disconnect to and from the enclave network idempotent (#2004)
## Description:
This PR makes the reverse proxy network calls idempotent so the engine
can be started when networking is already set up. This can happen when
you restart the docker daemon.

In addition, this PR appends the engine UUID to the reverse proxy
container name to make it unique across runs. This is an approach
similar to the engine naming scheme which is cleaner and safer.

## Is this change user facing?
NO

## References (if applicable):
#2005