fix: emui build (#1826)
## Description:
This PR brings the emui into the ci build/release process. It is a
reversion of last nights
https://github.com/kurtosis-tech/kurtosis/pull/1813.

Yesterday the build/release change for emui was abandoned because after
several attempts the publish step was not working. In
https://github.com/kurtosis-tech/kurtosis/commit/4ed9328ab2743c045a7dfa71c8c4e7d4122aae35
I have demonstrated that this pr works by:
* Removing the body of the publish jobs and replacing them with echo
statements
* Triggering publish jobs on `bgazzard-*` tags
* Pushing a tag `bgazzard-0.0.1` (which I have now deleted) to trigger
the `publish` workflow
* The publish workflow can be seen here
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/9407/workflows/ca6cef3b-4f71-481a-ab08-8787002ed9d0
and demonstrates that the `build_enclave_manager_webapp` job is now
successfully run as a dependency of `publish_engine_server_image`.

The existing build/test path which was approved previously is unchanged
by this pr and can be seen working in this workflow
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/9408/workflows/bb48e469-2ff7-433e-8da9-ad32620b0bbc.

## Is this change user facing?
NO