fix: emui enable exact package name searching (#2052)
## Description:
This PR implements exact package name searching in the EMUI. This allows
indexing errors to be surfaced in the UI, for example:


[exact-match.webm](https://github.com/kurtosis-tech/kurtosis/assets/4419574/8cf29524-59de-44fc-834b-c07852406a12)

This PR also fixes #2051 by treating arguments with no argtype (as is
the case for the `args` argument) as a `JSON` argument. This behaviour
for Form->kurtosis format is now consistent with the existing kurtosis
format->Form transformer.

## Is this change user facing?
YES

## References (if applicable):
- Fixes https://github.com/kurtosis-tech/kurtosis/issues/2051