fix: emui remove extra args transformation (#2063)
## Description:
This fixes an issue reported by @adschwartz and @chunha-park where
packages which use the catch all `args` arg, like the ethereum package
show the following error:


![image](https://github.com/kurtosis-tech/kurtosis/assets/4419574/15ef556d-3e29-46bf-ad5b-ee0f8fc24694)

This error is occuring because the `args` arg has no 'TopLevelType', so
prior to #2052 the `args` field was not parsed as json/yaml when read.
However in #2052 the default behaviour was modified to parse fields with
an unknown type as json/yaml. This created this new bug, since the
`args` field was no parsed twice.

This fix removes the additional parsing in the 'submit' flow.

## Is this change user facing?
YES

## References (if applicable):
-
https://sageroomworkspace.slack.com/archives/C060Z6HDYR3/p1705615425847449