chore: Increase the cypress default command timeout to 20s (#2068)
## Description:
This PR increases the default command timeout in cypress in an attempt
to address flakey tests reported by @h4ck3rk3y on slack.

Examples of flakey runs:
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10765/workflows/2d6e411a-d6b9-472b-9495-1ea69b304402/jobs/154148
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10773/workflows/4a2b122f-bdda-48f1-b4f3-df29ae021949/jobs/154294
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10755/workflows/f9e01095-678e-498c-94b9-ce47bfe771bd/jobs/154037
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10751/workflows/d9757f39-b633-4996-81a5-462cb21ebc5c/jobs/153944

Looking at these it seems another common thread is that the enclave list
test is flakey at an unknown point - specifically some promise exception
can be thrown. I think this is probably in the log viewer - and not
actually indicative of the test failing. For this reason I've disabled
the unhandled promise exception failures in this pr.

## Is this change user facing?
NO

## References (if applicable):
*
https://docs.cypress.io/guides/core-concepts/retry-ability#Increase-time-to-retry
*
https://docs.cypress.io/guides/references/configuration#Overriding-Options