feat: Enclave connect (#2117)
## Description:
We want the user to be able to connect from their local machine to an
enclave with a start from the UI. A connect button is added to the
enclave page with the CLI commands to paste. A CLI enclave connect
command is added.

## Is this change user facing?
YES

## References (if applicable):

(https://www.notion.so/kurtosistech/Connect-local-CLI-to-running-enclave-workflow-for-cloud-based-enclaves-via-frontend-636d41ca937c4e49b54161aa498704b1?pvs=4)