chore: emui enable cypress retries (#2081)
## Description:
@leoporoli has identified a case of the cypress tests again failing due
to a flakey test, subsequent to the attempt to fix in #2068.

Related failed runs are:
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10835/workflows/4982925a-fb9d-4153-97d1-d5ec7288bb36/jobs/155379
*
https://app.circleci.com/pipelines/github/kurtosis-tech/kurtosis/10835/workflows/d6944842-bde1-45dc-b438-472cf735e13d/jobs/155457

These jobs have both failed in the `Can create and update an enclave`
test, which is the most important ete one for the emui. On subsquent
retries @leoporoli was able to have the test succeed.

It's not clear to me what is causing this flakiness, so this pr makes
two changes:
* Increase some hardcoded timeouts which post #2068 are actually less
generous than the default (I think this may have been the issue on the
second failure)
* Allow up to 3 attempts on a test when running in CI.

## Is this change user facing?
NO

## References (if applicable):
*
https://sageroomworkspace.slack.com/archives/C05DZ4UABDZ/p1706185246883059