feat: kurtosis run `--non-blocking-tasks` (#2153)
## Description:
Adds a flag `--non-blocking-tasks`. If set, task services from `run_sh`
and `run_python` are not removed before continuing with execution. The
effects is small over Docker, but this significantly reduces run times
over k8s because of the overhead of removing a pod (up to 30s).

For the `ethereum-package` over k8s with images cached, running with the
flag goes from 4min to 1min. However, task services will have to be
removed manually after execution or via a `kurtosis clean`.

## Is this change user facing?
YES

## References:
https://github.com/kurtosis-tech/kurtosis/issues/2140