feat: emui builder python nodes (#2170)
## Description:
This PR primarily adds support for `run_python` nodes in the enclave
builder graph.

It removes the `exec` node that was added yesterday, as getting this
node correct in the topological sort was more complex than just merging
it's behaviour with the service node (so that's what I've done).

A couple of small changes are also include:
* memo behaviour on node selection was wrong,
* When zooming out show a simplified version of nodes
* Update icons and add some node type text to make it clearer what's
what.

## Is this change user facing?
YES (experimental flag)

## References (if applicable):
* discussed on slack