chore: Update k8s image pull error message to include a possible platform mismatch (#2209)
## Description:
K8S might not be able to pull the image if there is a typo in the name,
if the registry is not accessible but also if the image's platform is
not compatible with the one K8S is running on. We update the error
message to indicate that possibility.

## Is this change user facing?
YES

## References (if applicable):
Closes #2097

---------

Co-authored-by: Gyanendra Mishra <anomaly.the@gmail.com>