feat: adding enclave production mode value in the APIC primitive (#2407)
## Description
The `production mode` value is already in the `enclave` primitive but we
also want to add it in the `APIC` primitive this way we will be able to
get this data just by calling the `kurtosisBackend.GetAPIContainers()`
instead of having to call the `kurtosisBackend.GetEnclaves()`which will
demand more internal backend calls, and also because this data is stored
in the APIC´s env vars.

Adding it only in the Docker backend because production mode is not
implemented in Kubernetes yet.

## REMINDER: Tag Reviewers, so they get notified to review

## Is this change user facing?
NO

## References (if applicable)
This is part of the "Upgrade Kurtosis from the Cloud UI project"