fix: clarify error message for kurtosis portal command (#2433)
## Description
In certain cases, the kurtosis portal may not be able to reach your
cloud instance. In this case a `kurtosis run` command may output an
error that suggests checking the health of your kurtosis portal

```
[...]
Make sure Kurtosis Portal is running locally with 'kurtosis portal status' and
potentially 'kurtosis portal start'. If it is, make sure the remote server is
running and healthy
```

Running `kurtosis portal status` results in a slightly misleading error
message, for example:

```
Kurtosis Portal process is on PID 4615 but it is unreachable
```

`it is unreachable` suggests there is an issue with with `kurtosis
portal` process when in reality there is an issue with the cloud
instance being unreachable. This PR improves the error message.

## Is this change user facing?
YES

## References (if applicable)
- Slack discussion with @laurentluce