feat: Self cloud hosting (#2485)
## Description
The Kurtosis CLI supports the local use case (localhost) and the
Kurtosis cloud use case (cloud.kurtosis.com). We need to be able to
build the CLI once and deploy it to multiple environments with different
domain names. Add option to the engine to set a domain name used by the
EM UI. The domain name is set in a static js file read by the EM UI.
More details on this approach discussed here:
https://kurtosistech.slack.com/archives/C032ZB39AHH/p1717770540781579

Adding Nginx in front of the engine + install script will be done as a
cloud-config PR.

## Is this change user facing?
YES