fix: add suffix to fluentbit configurator
We run Kurtosis in our CI pipeline. Multiple processes create Kurtosis enclaves concurrenctly. When the timing is just right, it can lead to a race condition where the fluentbit configurator container sticks around and no further enclaves can be created until we manually delete the fluentbit container.

This PR adds a suffix to the configurator container so that it is isolated from other concurrent runs.