fix: add suffix to fluentbit configurator (#2567)
## Description

We run Kurtosis in our CI pipeline. Multiple processes create Kurtosis
enclaves concurrenctly. When the timing is just right, it can lead to a
race condition where the fluentbit configurator container sticks around
and no further enclaves can be created until we manually delete the
fluentbit container.

This PR adds a suffix to the configurator container so that it is
isolated from other concurrent runs.

## Is this change user facing?

No

Co-authored-by: Tedi Mitiku <tedi.m52@gmail.com>