feat!: impl dependency list flag  (#2543)
## Description
`kurtosis run github.com/kurtosis-tech/... --dependencies` returns a
`yaml` with a list of images and packages the run will need to depend
on.

Adding `--pull` will pull all the dependencies locally. The
`kurtosis.yml` will be updated to use local versions of packages and
Docker images are pulled onto the machine locally

## Is this change user facing?
YES