fix: use ISO 8601 year number (#2621)
## Description
The Vector logs aggregator output `%Y` specifier for year which is the
[proleptic Gregorian
year](https://docs.rs/chrono/0.4.19/chrono/format/strftime/index.html#specifiers)
instead of `%G` specifier the ISO 8601. Logs are read with the ISO
format so this created a mismatch in logs being read / written at the
end of 2024 (eg. `Y` output the year 2024 on Dec. 30th 2024 whereas the
ISO 8601 year on that date is 2025)

## Is this change user facing?
NO