feat: add user flag to service exec command (#2626)
## Description
We depend on some containers that use non-root users for certain things,
so would be nice to be able to `service exec --user foo` as those users,
similar as the docker CLI interface.

This PR adds --user to `service exec` and passes the user on the
RunExecCommand in the docker and kubernetes managers. I could not find a
good way to do it for kubernetes, but that might be appropriate for more
production like configurations.

## REMINDER: 
@tedim52

## Is this change user facing?
YES