feat: fluent bit logs collector in k8s backend (#2653)
## Description
This PR is part 1 in adding centralized logging to the K8s backend.
Similar to the Docker backend, this change implements the
deployment/management of a [fluentbit](https://docs.fluentbit.io/manual)
logs collector's via a k8s `DaemonSet`. The `DaemonSet` deploys fluent
bit log collection pods on every node in a k8s cluster. The logs
collector is configured to pick up logs from all user services in
enclaves running on that node.

Currently, it's configured to output the log stream to stdout - in pt 2.
the logs will be sent to a Vector aggregator.

## Is this change user facing?
NO