feat: export service logs (Docker) (#2668)
## Description
The following PR allows users to define custom Vector sinks for the
Docker backend.
It also updates vector to v0.45.0 and adds health check to the logs
aggregator.

Sink configurations will reside under the `logs-aggregator` key under
each cluster config. They are passed almost as-is to Vector, with the
exception of the `inputs` field, which we set as `fluent_bit`.
Validation of such configurations are also carried out by Vector.

An example configuration to allow logs export to AWS Opensearch is as
follows:

```
config-version: 3
should-send-metrics: true
kurtosis-clusters:
  docker:
    type: "docker"
    logs-aggregator: 
      sinks: 
        elasticsearch:
          type: "elasticsearch"
          opensearch_service_type: "serverless"
          bulk: 
            index: "kt-{{ enclave_uuid }}-{{ service_uuid }}"
          aws:
            region: "eu-west-1"
          auth:
            strategy: "aws"
            access_key_id: "<access_key_id>"
            secret_access_key: "<secret_access_key>"
          endpoints:
            - "<opensearch_endpoint>"
```

## Is this change user facing?
YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable)
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->
<!-- You can auto-close issues by putting "Fixes #XXXX" here. -->

---------

Co-authored-by: Tedi Mitiku <tedi.m52@gmail.com>