feat: vector logs aggregator in k8s (#2669)
## Description
This PR adds a vector logs aggregator to the kubernetes backend and has
it output logs to a shared volume, mouned to the engine pod.

Additionally, it migrates the k8s backend to use the
`PersistentVolumeLogsDatabaseClient` so now both Docker and K8s are
streaming from a persistent volume containing logs aggregated by Vector.

To use kubernetes logging, users will need to use the v4 config and
specify the engine node they'd like the engine to be placed on. Both the
aggregator and engine will be placed on this node every `kurtosis engine
start` so the engine has access to the same logs db.
```
config-version: 4
should-send-metrics: true
kurtosis-clusters:
  ...
  minikube:
    type: "kubernetes"
    config:
      kubernetes-cluster-name: "minikube"
      enclave-size-in-megabytes: 10
      storage-class: "standard"
      engine-node-name: "minikube-m02"
  ...
```

## Is this change user facing?
YES

## References
https://github.com/kurtosis-tech/kurtosis/pull/2653
https://github.com/kurtosis-tech/kurtosis/pull/2663