fix: docker auth wasn't being used by engine and API (#2699)
## Description
<!-- Describe this change, how it works, and the motivation behind it.
-->

There is a bug in the docker config storage creator which was introduced
in here https://github.com/kurtosis-tech/kurtosis/pull/2579 . This bug
causes no auth to ever be used by the API and Engine containers.
This PR fixes that and also adds some more tests for edge cases for repo
addresses that can exist in the config.json.

```diff
- if err != nil && creds != nil {
+ if err == nil && creds != nil {
```

## REMINDER: Tag Reviewers, so they get notified to review

## Is this change user facing?
NO
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable)
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->
<!-- You can auto-close issues by putting "Fixes #XXXX" here. -->