fix: give grafana root access (#2706)
## Description
A user reported the grafana container not having access to the
bindmounted datasource yaml within the container. This was due to
varying permissions from the original tmp file created by the CLI and
inherited by the bindmounted file. To get around this, this change gives
the grafana container root user which fixed their issue.

## Is this change user facing?
YES

### References

https://github.com/kurtosis-tech/kurtosis/pull/2696#issuecomment-2789352402