fix: init remove funcs (#2734)
## Description
`var removeLogsAggregator/CollectorFuns func()` was initializing the
remove function as `nil`.

In the case where `logsAggregator/Collector.CreateAndStart` fails or if
the `get...ResourcesForCluster` call fails, the remove function never
gets initialized. When the error returns, the defer will attempt to
execute a nil function which causes a segfault.

Initializing the function to a no op solves this issue.

## Is this change user facing?
YES

## References
https://github.com/kurtosis-tech/kurtosis/issues/2704

https://discord.com/channels/783719264308953108/1364854129310044243/1364978226757107712