ci: fix flaky test `TestStartosis_AddServiceWithReadyConditionsCheckFail` (#2762)
## Description
`TestStartosis_AddServiceWithReadyConditionsCheckFail` occasionally
fails in CI:
![Screenshot 2025-06-18 at 4 15
44 PM](https://github.com/user-attachments/assets/35ab2201-2ba7-4896-9a15-5ff7b6705411)



https://github.com/kurtosis-tech/kurtosis/blob/fccb53c2e41043164a45db856d85bfaf4847dd12/core/server/api_container/server/startosis_engine/kurtosis_instruction/shared_helpers/service_helpers.go#L134

`ctxWithDeadline` and `interruptChan` are set to the same `timeout`. A
race condition was happening where if `ctxWithDeadline` timed out before
`interruptChan` signaled, the `execFunc` would get called -> fail with
`context deadline exceeded` making the last error a `recipeErr`. When
`interruptChan` eventually signaled, the loop would exit with that
`recipeErr` (the one displayed in the pic) as opposed to the expected
`assertErr`.

To fix this, I added a small buffer to `ctxWithDeadline` timeout so
`interruptChan` triggers before context deadline is exceeded. Not too
sure how both are being used to track timeouts or why `ctxWithDeadline`
is used instead of `ctxWithCancel` but much of that context is in this
PR: https://github.com/kurtosis-tech/kurtosis/pull/480

## Is this change user facing?
NO